# app/spec/swagger.yaml
openapi: 3.0.0
info:
  title: flowerstand
  contact:
    name: Tomoyuki Sato
    url: https://mstdn.schoolidol.club/@yazin
  license:
    name: MIT
    url: http://opensource.org/licenses/MIT
  version: 1.0.0
servers:
- url: https://flowerstand.schoolidol.club/api
paths:
  /flowerstands:
    get:
      summary: searches flower stands. returns all flower stands if no parameter specified.
      parameters:
      - name: eventId
        in: query
        required: false
        style: form
        explode: true
        schema:
          type: integer
      - name: baseDesignId
        in: query
        required: false
        style: form
        explode: true
        schema:
          type: integer
      - name: groupId
        in: query
        required: false
        style: form
        explode: true
        schema:
          type: integer
      - name: offset
        in: query
        required: false
        style: form
        explode: true
        schema:
          type: integer
      - name: limit
        in: query
        required: false
        style: form
        explode: true
        schema:
          type: integer
      - name: showPastEvents
        in: query
        required: false
        style: form
        explode: true
        schema:
          type: integer
          description: 0 if limit search result to present and future event, 1 otherwise.
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/FlowerStand'
        "400":
          description: one or more parameters are not valid.
    post:
      summary: creates a new flower stand.
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/body'
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FlowerStandWithKeys'
        "204":
          description: baseDesignId or eventId does not exists.
        "400":
          description: possibly posted image contains sensitive contents.
  /flowerstands/verify:
    post:
      summary: verify admin key.
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/body_1'
        required: true
      responses:
        "200":
          description: OK
        "204":
          description: flower stand with specified id does not exists.
        "401":
          description: invalid adminKey.
  /flowerstands/preview:
    post:
      summary: returns a preview image.
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/body_2'
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/inline_response_200'
        "204":
          description: baseDesignId or eventId does not exists.
        "400":
          description: possibly posted image contains sensitive contents.
  /flowerstands/{id}:
    get:
      summary: returns a flower stand.
      parameters:
      - name: id
        in: path
        description: id of flower stand.
        required: true
        style: simple
        explode: false
        schema:
          type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FlowerStand'
        "204":
          description: flower stand with specified id does not exists.
    put:
      summary: updates a flower stand.
      parameters:
      - name: id
        in: path
        description: id of flower stand to update.
        required: true
        style: simple
        explode: false
        schema:
          type: integer
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/body_3'
        required: true
      responses:
        "200":
          description: OK
        "204":
          description: flower stand with specified id does not exists.
        "401":
          description: invalid adminKey.
    delete:
      summary: deletes a flower stand.
      parameters:
      - name: id
        in: path
        description: id of flower stand to delete.
        required: true
        style: simple
        explode: false
        schema:
          type: integer
      - name: adminKey
        in: query
        description: administration key issued on create flower stand.
        required: true
        style: form
        explode: true
        schema:
          type: string
      responses:
        "200":
          description: OK
        "204":
          description: FlowerStand of specified id does not exists.
        "401":
          description: invalid adminKey.
...
components:
  schemas:
    FlowerStand:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        presentTo:
          type: string
        presentFrom:
          type: string
        event:
          $ref: '#/components/schemas/Event'
        description:
          type: string
        projectUrl:
          type: string
        organizerName:
          type: string
        participants:
          type: array
          items:
            $ref: '#/components/schemas/Participant'
        baseDesign:
          $ref: '#/components/schemas/BaseDesign'
        imageUrl:
          type: string
    FlowerStandWithKeys:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        presentTo:
          type: string
        presentFrom:
          type: string
        event:
          $ref: '#/components/schemas/Event'
        description:
          type: string
        projectUrl:
          type: string
        organizerName:
          type: string
        participants:
          type: array
          items:
            $ref: '#/components/schemas/Participant'
        baseDesign:
          $ref: '#/components/schemas/BaseDesign'
        imageUrl:
          type: string
        adminKey:
          type: string
        participationCode:
          type: string
    Participant:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        twitterId:
          type: string
        instagramId:
          type: string
        facebookId:
          type: string
        mastodonId:
          type: string
    BaseDesign:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        imageUrl:
          type: string
        sortOrder:
          type: integer
        group:
          $ref: '#/components/schemas/Group'
    Event:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        startDate:
          type: string
        endDate:
          type: string
        groups:
          type: array
          items:
            $ref: '#/components/schemas/Group'
    Group:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
    body:
      required:
      - baseDesignId
      - eventId
      - name
      - organizerName
      - presentFrom
      - presentTo
      type: object
      properties:
        name:
          type: string
        presentTo:
          type: string
        presentFrom:
          type: string
        organizerName:
          type: string
        description:
          type: string
        projectUrl:
          type: string
        baseDesignId:
          type: integer
        eventId:
          type: integer
        panel:
          type: string
          description: optional illust panel image (jpeg or png), Data URI format.
    body_1:
      required:
      - adminKey
      - flowerStandId
      type: object
      properties:
        flowerStandId:
          type: integer
          description: flower stand id to verify.
        adminKey:
          type: string
          description: adminKey to verify.
    body_2:
      required:
      - baseDesignId
      - presentFrom
      - presentTo
      type: object
      properties:
        presentTo:
          type: string
        presentFrom:
          type: string
        baseDesignId:
          type: number
        eventId:
          type: number
        panel:
          type: string
          description: optional illust panel image (jpeg or png), Data URI format.
    inline_response_200:
      type: object
      properties:
        imageUrl:
          type: string
          description: URL of preview image.
    body_3:
      required:
      - adminKey
      type: object
      properties:
        adminKey:
          type: string
          description: administration key issued on create flower stand.
        name:
          type: string
          description: name of flower stand.
        presentTo:
          type: string
          description: recipient name to be displayed on flower stand.
        presentFrom:
          type: string
          description: presenter name to be displayed on flower stand.
        organizerName:
          type: string
          description: organizer name of flower stand. this is not displayed on flower
            stand.
        description:
          type: string
          description: description of flower stand.
        projectUrl:
          type: string
          description: URL of flower stand project.
...